+++ pp/samples/18_port_markers.py
@@ -2,17 +2,17 @@
 
 """
 import pp
-from pp.add_pins import _add_pins_triangle
+from pp.add_pins import add_pins_triangle
 from pp.component import Component
 
 
 def test_pins_custom() -> Component:
-    """We can even define the `pins_function` that we use to add markers to each port"""
+    """You can define the `pins_function` that we use to add markers to each port"""
     c = pp.c.waveguide()
-    cc = pp.add_pins(component=c, function=_add_pins_triangle)
-    return cc
+    add_pins_triangle(component=c)
+    return c
 
 
 if __name__ == "__main__":
     c = test_pins_custom()
-    c.show()
+    c.show(show_ports=False)